requirements.txt
# AI Knowledge Assistant - Production Dependencies

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI & ML - Ollama (Local via HTTP)
# Using requests library for Ollama HTTP API
langchain==0.1.4
langchain-community==0.0.16

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.3

# PDF Processing
PyPDF2==3.0.1

# Redis Caching
redis==5.0.1

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
